requirements.txt
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6

cmake==3.26.4
conda-pack==0.7.1
contourpy==1.1.0
cox==0.1.post3
cycler==0.11.0
dill==0.3.7
docker-pycreds==0.4.0
easydict==1.11
executing==1.2.0
filelock==3.12.2
fonttools==4.42.1
fsspec==2023.6.0
ftfy==6.1.1
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
gpustat==1.1
grpcio==1.59.3
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.1
ipdb==0.13.13
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.5
lit==16.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
open-clip-torch==2.23.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==10.0.0
py3nvml==0.2.7
Pygments==2.15.1
pyparsing==3.0.9
PySocks==1.7.1
pytorch-fid==0.3.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.99
sentry-sdk==1.29.2
setGPU==0.0.7
setproctitle==1.3.2
smmap==5.0.0
soupsieve==2.5
sympy==1.12
tensorboardX==2.6.2.2
timm==0.9.2
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
wandb==0.15.8
wcwidth==0.2.6
xmltodict==0.13.0
zipp==3.16.2
